Moved ingest errors to errors namespace
