+++ app/lib/dmao/ingesters/errors/ingest_file_error.rb
@@ -1,0 +1,17 @@
+module DMAO
+  module Ingesters
+    module Errors
+
+      class IngestFileError < IngestError
+
+        def initialize(msg="Error finding file", validation_errors=nil)
+
+          super(msg, validation_errors)
+
+        end
+
+      end
+
+    end
+  end
+end
