--- app/lib/dmao/ingesters/ingest_error.rb
@@ -1,19 +1,0 @@
-module DMAO
-  module Ingesters
-
-    class IngestError < StandardError
-
-      attr_reader :validation_errors
-
-      def initialize(msg="Error ingesting object to system", validation_errors=nil)
-
-        @validation_errors = validation_errors
-
-        super(msg)
-
-      end
-
-    end
-
-  end
-end
