--- app/lib/dmao/ingesters/ingest_file_error.rb
@@ -1,15 +1,0 @@
-module DMAO
-  module Ingesters
-
-    class IngestFileError < IngestError
-
-      def initialize(msg="Error finding file", validation_errors=nil)
-
-        super(msg, validation_errors)
-
-      end
-
-    end
-
-  end
-end
